# schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "AudioConfig-Output": {
        "additionalProperties": false,
        "properties": {
          "gains_db": {
            "additionalProperties": {
              "type": "number"
            },
            "propertyNames": {
              "format": "uuid"
            },
            "title": "Gains Db",
            "type": "object"
          },
          "mode": {
            "$ref": "#/components/schemas/AudioMode",
            "default": "mix"
          },
          "single_clip_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Single Clip Id"
          }
        },
        "required": [
          "mode",
          "gains_db",
          "single_clip_id"
        ],
        "title": "AudioConfig",
        "type": "object"
      },
      "AudioMode": {
        "enum": [
          "mix",
          "single"
        ],
        "title": "AudioMode",
        "type": "string"
      },
      "ClipCreate": {
        "additionalProperties": false,
        "properties": {
          "path": {
            "description": "Absolute path; the file is referenced in place",
            "minLength": 1,
            "title": "Path",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/ClipRole",
            "default": "speaker"
          },
          "speaker_label": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker Label"
          }
        },
        "required": [
          "path"
        ],
        "title": "ClipCreate",
        "type": "object"
      },
      "ClipOut": {
        "additionalProperties": false,
        "properties": {
          "file_status": {
            "$ref": "#/components/schemas/FileStatus"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_reference": {
            "title": "Is Reference",
            "type": "boolean"
          },
          "media": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MediaInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/ClipRole"
          },
          "speaker_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker Label"
          },
          "sync": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SyncResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "project_id",
          "path",
          "name",
          "role",
          "speaker_label",
          "media",
          "sync",
          "file_status",
          "is_reference"
        ],
        "title": "ClipOut",
        "type": "object"
      },
      "ClipRole": {
        "enum": [
          "speaker",
          "wide",
          "broll"
        ],
        "title": "ClipRole",
        "type": "string"
      },
      "ClipUpdate": {
        "additionalProperties": false,
        "properties": {
          "force": {
            "default": false,
            "description": "Relink even if the new file looks different",
            "title": "Force",
            "type": "boolean"
          },
          "path": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Relink a moved file",
            "title": "Path"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClipRole"
              },
              {
                "type": "null"
              }
            ]
          },
          "speaker_label": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker Label"
          }
        },
        "title": "ClipUpdate",
        "type": "object"
      },
      "CutList-Input": {
        "additionalProperties": false,
        "properties": {
          "audio": {
            "$ref": "#/components/schemas/AudioConfig-Input"
          },
          "fps": {
            "$ref": "#/components/schemas/Rational"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "removals": {
            "items": {
              "$ref": "#/components/schemas/Removal-Input"
            },
            "title": "Removals",
            "type": "array"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/Segment-Input"
            },
            "minItems": 1,
            "title": "Segments",
            "type": "array"
          },
          "version": {
            "default": 1,
            "description": "Edit revision; bumps on every change",
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "project_id",
          "fps",
          "segments"
        ],
        "title": "CutList",
        "type": "object"
      },
      "CutList-Output": {
        "additionalProperties": false,
        "properties": {
          "audio": {
            "$ref": "#/components/schemas/AudioConfig-Output"
          },
          "fps": {
            "$ref": "#/components/schemas/Rational"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "removals": {
            "items": {
              "$ref": "#/components/schemas/Removal-Output"
            },
            "title": "Removals",
            "type": "array"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/Segment-Output"
            },
            "minItems": 1,
            "title": "Segments",
            "type": "array"
          },
          "version": {
            "default": 1,
            "description": "Edit revision; bumps on every change",
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "schema_version",
          "version",
          "project_id",
          "fps",
          "segments",
          "removals",
          "audio"
        ],
        "title": "CutList",
        "type": "object"
      },
      "CutListIn": {
        "additionalProperties": false,
        "properties": {
          "cutlist": {
            "$ref": "#/components/schemas/CutList-Input"
          }
        },
        "required": [
          "cutlist"
        ],
        "title": "CutListIn",
        "type": "object"
      },
      "CutListOut": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "cutlist": {
            "$ref": "#/components/schemas/CutList-Output"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "source",
          "created_at",
          "cutlist"
        ],
        "title": "CutListOut",
        "type": "object"
      },
      "CutListVersion": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "segments": {
            "title": "Segments",
            "type": "integer"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "source",
          "segments",
          "created_at"
        ],
        "title": "CutListVersion",
        "type": "object"
      },
...
      "FileStatus": {
        "enum": [
          "ok",
          "missing",
          "changed",
          "unknown"
        ],
        "title": "FileStatus",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MediaInfo": {
        "additionalProperties": false,
        "description": "Facts about a source file, read with ffprobe (Phase 1).",
        "properties": {
          "audio_channels": {
            "default": 0,
            "minimum": 0.0,
            "title": "Audio Channels",
            "type": "integer"
          },
          "audio_codec": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Codec"
          },
          "audio_sample_rate": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Sample Rate"
          },
          "duration_frames": {
            "minimum": 0.0,
            "title": "Duration Frames",
            "type": "integer"
          },
          "fps": {
            "$ref": "#/components/schemas/Rational"
          },
          "height": {
            "exclusiveMinimum": 0.0,
            "title": "Height",
            "type": "integer"
          },
          "is_vfr": {
            "title": "Is Vfr",
            "type": "boolean"
          },
          "video_codec": {
            "title": "Video Codec",
            "type": "string"
          },
          "width": {
            "exclusiveMinimum": 0.0,
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "fps",
          "is_vfr",
          "duration_frames",
          "width",
          "height",
          "video_codec",
          "audio_codec",
          "audio_sample_rate",
          "audio_channels"
        ],
        "title": "MediaInfo",
        "type": "object"
      },
...
      "Rational": {
        "additionalProperties": false,
        "description": "An exact positive rational number, e.g. a frame rate of 30000/1001.",
        "properties": {
          "den": {
            "exclusiveMinimum": 0.0,
            "title": "Den",
            "type": "integer"
          },
          "num": {
            "exclusiveMinimum": 0.0,
            "title": "Num",
            "type": "integer"
          }
        },
        "required": [
          "num",
          "den"
        ],
        "title": "Rational",
        "type": "object"
      },
...
      "Removal-Output": {
        "additionalProperties": false,
        "properties": {
          "approved": {
            "default": false,
            "title": "Approved",
            "type": "boolean"
          },
          "end_frame": {
            "description": "Exclusive",
            "exclusiveMinimum": 0.0,
            "title": "End Frame",
            "type": "integer"
          },
          "kind": {
            "$ref": "#/components/schemas/RemovalKind"
          },
          "start_frame": {
            "minimum": 0.0,
            "title": "Start Frame",
            "type": "integer"
          }
        },
        "required": [
          "start_frame",
          "end_frame",
          "kind",
          "approved"
        ],
        "title": "Removal",
        "type": "object"
      },
      "RemovalKind": {
        "enum": [
          "filler",
          "silence",
          "manual"
        ],
        "title": "RemovalKind",
        "type": "string"
      },
...
      "SyncResult": {
        "additionalProperties": false,
        "description": "How a clip lines up with the project's reference clip.\n\nSign convention (shared with ``benchmark.GroundTruth``):\n    ``offset_samples`` = position of an event in THIS clip\n                         minus position of the same event in the REFERENCE clip.\n    Positive  -> this clip started recording EARLIER than the reference.\n    Negative  -> this clip started recording LATER than the reference.\n\n``drift_ppm``: how much faster this clip's clock runs than the reference,\nin parts per million. A value of +100 means the clip accumulates 0.36 s\nof extra duration per hour. This is a model parameter, not a timeline\nposition, so a float is acceptable here.",
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "drift_ppm": {
            "default": 0.0,
            "title": "Drift Ppm",
            "type": "number"
          },
          "offset_samples": {
            "title": "Offset Samples",
            "type": "integer"
          },
          "reference_clip_id": {
            "format": "uuid",
            "title": "Reference Clip Id",
            "type": "string"
          },
          "sample_rate": {
            "exclusiveMinimum": 0.0,
            "title": "Sample Rate",
            "type": "integer"
          }
        },
        "required": [
          "reference_clip_id",
          "offset_samples",
          "sample_rate",
          "drift_ppm",
          "confidence"
        ],
        "title": "SyncResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Local backend: projects, clips, jobs (sync, auto-edit, render).",
    "title": "Multicam Studio API",
    "version": "0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/clips/{clip_id}": {
      "delete": {
        "operationId": "delete_clip_api_clips__clip_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "clip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Clip Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Clip",
        "tags": [
          "clips"
        ]
      },
      "get": {
        "operationId": "get_clip_api_clips__clip_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "clip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Clip Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClipOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Clip",
        "tags": [
          "clips"
        ]
      },
      "patch": {
        "operationId": "update_clip_api_clips__clip_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "clip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Clip Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClipUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClipOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Clip",
        "tags": [
          "clips"
        ]
      }
    },
...
    "/api/projects/{project_id}/clips": {
      "get": {
        "operationId": "list_clips_api_projects__project_id__clips_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClipOut"
                  },
                  "title": "Response List Clips Api Projects  Project Id  Clips Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Clips",
        "tags": [
          "clips"
        ]
      },
      "post": {
        "operationId": "add_clip_api_projects__project_id__clips_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClipCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClipOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Clip",
        "tags": [
          "clips"
        ]
      }
    },
    "/api/projects/{project_id}/cutlist": {
      "get": {
        "operationId": "get_cutlist_api_projects__project_id__cutlist_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CutListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Cutlist",
        "tags": [
          "cutlist"
        ]
      },
      "put": {
        "description": "Save an edited cutlist as a new version (the engine validates it first).",
        "operationId": "save_cutlist_api_projects__project_id__cutlist_put",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CutListIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CutListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Cutlist",
        "tags": [
          "cutlist"
        ]
      }
    },
    "/api/projects/{project_id}/cutlist/versions": {
      "get": {
        "operationId": "list_versions_api_projects__project_id__cutlist_versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CutListVersion"
                  },
                  "title": "Response List Versions Api Projects  Project Id  Cutlist Versions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Versions",
        "tags": [
          "cutlist"
        ]
      }
    },
    "/api/projects/{project_id}/cutlist/versions/{version}": {
      "get": {
        "operationId": "get_version_api_projects__project_id__cutlist_versions__version__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CutListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Version",
        "tags": [
          "cutlist"
        ]
      }
    },
...
  }
}